f32

+++ tinygrad/idk_torch.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 torch.backends.cudnn.deterministic = True
 
 # Assuming input tensor shape is (batch_size, channels, depth, height, width)
-input_tensor = torch.randn((1, 32, 128, 128, 128))
+input_tensor = torch.randn((1, 32, 128, 128, 128), dtype=torch.float32)
 
 # Define 3D convolution layer
 conv = nn.Conv3d(in_channels=32, out_channels=32, kernel_size=(3, 3, 3), padding=(1, 1, 1))
